Move scene pass types into pipeline mod
