+++ src/pipeline/render_pass.rs
@@ -1,5 +1,41 @@
 use crate::pipeline::Context;
+use crate::scene::SceneCore;
 use crate::{screen_quad, shader, DrawError};
+
+pub struct ShadowPass<C: SceneCore> {
+    pub program: glium::Program,
+
+    /// The transformed shader core that was used for building the `program`.
+    /// Currently this is basically just phantom data.
+    #[allow(dead_code)]
+    pub shader_core: shader::Core<(Context, C::Params), C::Instance, C::Vertex>,
+}
+
+#[derive(Debug, Clone)]
+pub struct ShadedScenePassSetup {
+    pub draw_shadowed: bool,
+    pub draw_glowing: bool,
+}
+
+pub struct ShadedScenePass<C: SceneCore> {
+    pub setup: ShadedScenePassSetup,
+
+    pub program: glium::Program,
+
+    /// The transformed shader core that was used for building the `program`.
+    /// Currently this is basically just phantom data.
+    #[allow(dead_code)]
+    pub shader_core: shader::Core<(Context, C::Params), C::Instance, C::Vertex>,
+}
+
+pub struct PlainScenePass<C: SceneCore> {
+    pub program: glium::Program,
+
+    /// The transformed shader core that was used for building the `program`.
+    /// Currently this is basically just phantom data.
+    #[allow(dead_code)]
+    pub shader_core: shader::Core<(Context, C::Params), C::Instance, C::Vertex>,
+}
 
 pub trait RenderPassComponent {
     fn clear_buffers<F: glium::backend::Facade>(&self, facade: &F) -> Result<(), DrawError>;
